requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
django-rename-app==0.1.2
futures==3.1.1
goslate==1.5.1
gunicorn==20.0.4
hashids==1.3.1
idna==2.10
lxml==4.6.2
psycopg2==2.8.6
pytz==2020.4
requests==2.25.0
shortuuid==1.0.1
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
djangorestframework==3.1.2
